requirements.txt
amqp==2.6.1
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.12.2
billiard==3.6.4.0
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.2
django-crispy-forms==1.11.2
django-environ==0.4.5
django-phonenumber-field==5.2.0
django-registration-redux==2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Faker==8.1.4
filelock==3.0.12
idna==2.10
kombu==4.6.11
lxml==4.9.2
numpy==1.24.3
pandas==2.0.2
phonenumbers==8.13.13
pipenv==2020.11.15
product==0.1.1.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-http-client==3.3.7
pytz==2021.1
requests==2.25.1
sendgrid==6.7.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.1
starkbank-ecdsa==2.2.0
text-unidecode==1.3
tzdata==2023.3
urllib3==1.26.4
vine==1.3.0
virtualenv==20.4.6
virtualenv-clone==0.5.4
xlrd==2.0.1
